\name{acf}
\alias{acf}
\title{
acf function
}
\description{
This calls on the acf function in the stats package.
}
\author{
Original: Paul Gilbert, Martyn Plummer. Extensive modifications and univariate case of pacf by B. D. Ripley.
}

\keyword{methods}
\keyword{univar}
\keyword{models}
\keyword{ts}
