\name{checkData}
\alias{checkData}
\title{
checkData
}
\description{
Checks that all data is numeric before data is passed in to \code{PELT.online} functions.

WARNING: Developer Use.
}
\usage{
checkData(data)
}
\arguments{
  \item{data}{The vector or matrix of data that is being checled}
}
\details{
	This function is used as a check.
}

\keyword{methods}
\keyword{univar}
\keyword{models}
\keyword{ts}
