\name{lastchange-internal}
\alias{compress_lastchange}
\alias{expand_lastchange}
\alias{lastchangecpts<-,ocpt,matrix-method}
\alias{lastchangecpts<-,ocpt,numeric-method}
\alias{lastchangecpts<-,ocpt.reg,matrix-method}
\alias{lastchangecpts<-,ocpt.reg,numeric-method}
\alias{lastchangelike<-,ocpt,matrix-method}
\alias{lastchangelike<-,ocpt,numeric-method}
\alias{lastchangelike<-,ocpt.reg,matrix-method}
\alias{lastchangelike<-,ocpt.reg,numeric-method}
\title{Internal helpers for last-change bookkeeping}
\description{
Internal helper functions and S4 replacement methods used by the online
changepoint algorithms to maintain last changepoint locations and likelihood
values. They are not intended to be called directly by users.
}
\keyword{internal}