\name{ocpts<-}
\alias{ocpts<-}
\title{
Generic Function - ocpts<-
}
\description{
Generic function
}
\usage{
ocpts(object)<-value
}
\arguments{
  \item{object}{
	Depending on the class of \code{object} depends on the method used (and if one exists)
}
  \item{value}{
	Replacement value
}
}
\details{
	Generic function.
}
\value{
	Depends on the class of \code{object}, see individual methods
}
\author{
Rebecca Killick
}


\seealso{
\code{\link{ocpts<--methods}}
}
\examples{
x=new("ocpt") # new cpt object
}

\keyword{methods}
\keyword{ocpt}
\keyword{internal} 